Support for using printv in a dumped core image
This adds support for using printv as part of a core image dumped with SBCL. The problem was that the default output stream is bound to a stream that's valid in the original process, but is closed as part of dumping the image, so no longer prints output in an SBCL using this dumped core image. This fixes that.